move Error objects and is_empty into misc_utils (prevent circular imports)
