+++ src/SOPRANO/misc_utils.py
@@ -1,0 +1,18 @@
+import pathlib
+
+
+def is_empty(path: pathlib.Path) -> bool:
+    """
+    Checks whether file at path has size of zero
+    :param path: pathlib Path object
+    :return: True if path is empty else False
+    """
+    return path.stat().st_size == 0
+
+
+class MissingDataError(Exception):
+    pass
+
+
+class SOPRANOError(Exception):
+    pass
